requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
feedparser==6.0.11
beautifulsoup4==4.12.2
anthropic==0.25.0
python-dotenv==1.0.0
apscheduler==3.10.4
